Add a way to convert single XML nodes in json arrays
Exemple:

    <grid><rows/><rows/></grid>

Renders as:

    {grid: { rows: [{}, {}] } }

But:

    <grid><rows/></grid>

Rendered, before this commit, as:

    {grid: { rows: {} } }

As there was only one node, JXON didn't know that we wanted an Array.

So I added a "rToArray" parameter to the "build" and "createObjTree"
JXON functions, which is a regular expression to check if a particular
XML node should always be converted as an Array.

I also increased the verbosity when calling "build" to have an empty
JSON object instead of "true" when a node as no children/attributes.