Attach/detach modules to cell when needed
A module is not directly rendered into its cell, to be able to easily
move it, for exampe in drag&drop mode, or when adding a new module.

Instead, they are rendered in "div" holders, not displayed on the dom,
and they are attached to the cell when the cell was just rendered, and
detached from the the cell when it will be rendered.

It avoids having to recompute the whole module when a cell changes.

PS:

- add "json-stable-stringify" module which is a deterministic version
of JSON.stringify, as we use a stringified version of the cell content
attributes (which will be the props of the module component) as a key
for the cache

- add "jasmine-react" to spy on react component methods (it's my fork,
to integrate a PR to support react 0.12, and to correct a bug)