Make drag action working, with strong optimizations
Each "module" cell will have no direct react children.

But, using the "NodesHolder" mixin, a module react component, attached
to its own root node to be strictly independant from the grid, will be
attached to the cell after mouting/updating, and detached before
unmounting/updating.

In design mode, the module is not anymore attached to the cell, but a
"module holder" takes its place, holding himself the module.

This module holder is also a react component attached to its own root
node, to stay alive even when the cell node is destroyed when entering
drag mode. It's the module holder that handle the drag operation, adding
a cover above the module to avoid accidental interactions with the
modules content.