Make grid drag and drop working
As we remove the dragged element from the dom, browsers are not able to
fire `dragend` event. So we have to manage it ourself, and to detect
drop on the whole document surface to cancel the dragand drop operation
at the store level.

Also, if the user is really fast, the `drop` event is not fired at all.
So in this case we rely on the `mousedown` or `mousemove` event that can
only be triggered when the drop already occured, to detect that it as
occured.