Resolve pbm with dragLeave on holder fired before fakedrop
When a fake drop is detected, it means that the drop operation is done,
but we don't have the drop event.
But the browser, in some cases, still fire the ondragleave event of the
moduleholder (on "hovering" mode), so the "stopHovering" operation is
called, just before the fakedrop is catched, so we don't have an
hovering anymore.

The only solution I found is to delay the call to "stopHovering" to let
the fakedrop happen before.