Grid mobile navigation
In mobile mode, all module cells are on the same line (via css).

Only one cell is visible on the screen at a time.

Going from one cell to another is done by dragging left/right, or fast
drag (swipe) for fast access.

There is an special effect when trying to go beyond the borders.