Make jsdoc works again
It's possible that this problem will disappear when using es6 classes,
but for now, jsdoc doesn't display doc for things inside method calls,
so we create a "fake" object, for the doc, then you pass it to react to
export it.
Also, maybe it won't work because jsdoc add `exports` in path in its
tree, so it's not usable.